+++ delairstack/__init__.py
@@ -3,6 +3,6 @@ from pkgutil import extend_path
 
 __all__ = ('DelairStackSDK', )
 
-__version__ = '1.7.6'
+__version__ = '1.7.7'
 
 __path__ = extend_path(__path__, __name__)